Optimizacion para no llamar metodos min y max consecutivamente

+++ Semana_1/Actividades/Verificador.java
@@ -7,24 +7,24 @@ public class Verificador
     
     public int verificarRectangulos(Rectangulo rectA, Rectangulo rectB) {
 
-        int rectA_x1 = (int) rectA.getMinX();
-        int rectA_y1 = (int) rectA.getMinY();
-        int rectA_x2 = (int) rectA.getMaxX();
-        int rectA_y2 = (int) rectA.getMaxY();
-        
-        if (rectA_x1 < rectA_x2() && rectA_x2 > rectA_x1 // si se sobreponen en el eje x
+        double rectA_x1 = rectA.getMinX();
+        double rectA_y1 = rectA.getMinY();
+        double rectA_x2 = rectA.getMaxX();
+        double rectA_y2 = rectA.getMaxY();
+
+        if (rectA_x1 < rectA_x2 && rectA_x2 > rectA_x1 // si se sobreponen en el eje x
                 && // Y
             rectA_y1 < rectA_y2 && rectA_y2 > rectA_y1) { // se sobreponen en el eje y
             return SOBREPONEN;
         }
             
-        else if ((rectA_x2 == rectA_x1 && !(rectA_y2 < rectA_y1 || rectA_y1 > rectA_y2)) || 
+        else if ((rectA_x2 == rectA_x1 && !(rectA_y2 < rectA_y1 || rectA_y1 > rectA_y2)) ||
                  (rectA_x1 == rectA_x2 && !(rectA_y2 < rectA_y1 || rectA_y1 > rectA_y2)) ||
                  (rectA_y2 == rectA_y1 && !(rectA_x2 < rectA_x1 || rectA_x1 > rectA_x2)) ||
                  (rectA_y1 == rectA_y2 && !(rectA_x2 < rectA_x1 || rectA_x1 > rectA_x2))){
             return JUNTAN;
         }
-        
+
         else // en caso no se sobrepongan ni se junten queda solo la opcion de que sean disjuntos
         {
             return DISJUNTOS;
